+++ fair-timetable/src/blocks/timetable/index.js
@@ -1,0 +1,4 @@
+/**
+ * Timetable block registration
+ */
+import "./style.css";